more slightly less bad code but still kinda bad

+++ src/arch/riscv64/kernel/mod.rs
@@ -106,7 +106,7 @@ pub fn get_mbinfo() -> usize {
 	unsafe { volatile_load(&BOOT_INFO.mb_info) as usize }
 }
 
-pub fn get_processor_count() -> usize {
+pub fn get_processor_count() -> u32 {
 	unsafe { volatile_load(&BOOT_INFO.cpu_online) as usize }
 }
